Replace ImagesList with SimpleList
