+++ web/src/components/SimpleList.tsx
@@ -6,6 +6,7 @@ export interface Props<T> {
   children?: ReactNode;
   entries: T[];
   entryName: (entry: T) => string;
+  entryPreviewUrl?: (entry: T) => string;
   onClickEntry: (entry: T) => void;
   searchPlaceholder: string;
 }
@@ -14,6 +15,7 @@ export function SimpleList<T>({
   children,
   entries,
   entryName,
+  entryPreviewUrl,
   onClickEntry,
   searchPlaceholder,
 }: Props<T>) {
@@ -34,6 +36,14 @@ export function SimpleList<T>({
       <div className="entries">
         {showEntries.map((entry, i) => (
           <div key={i} className="entry" onClick={() => onClickEntry(entry)}>
+            {entryPreviewUrl && (
+              <div
+                className="preview"
+                style={{
+                  backgroundImage: `url(${entryPreviewUrl(entry)})`,
+                }}
+              />
+            )}
             <span className="name">{entryName(entry)}</span>
           </div>
         ))}
